refactor: split notification store actions
